+++ src/lib/isEmail.js
@@ -25,11 +25,7 @@ export default function isEmail(str, options) {
   assertString(str);
   options = merge(options, default_email_options);
 
-  if (options.require_display_name) {
-    options.allow_display_name = true;
-  }
-
-  if (options.allow_display_name) {
+  if (options.require_display_name || options.allow_display_name) {
     const display_email = str.match(displayName);
     if (display_email) {
       str = display_email[1];